Better .vpst support, various renames

+++ Valestrom/Metal/src/net/verdagon/vale/metal/instructions.scala
@@ -81,7 +81,7 @@ case class ConstantStrH(
 // Produces a float.
 case class ConstantF64H(
   // The value of the float.
-  value: Float
+  value: Double
 ) extends ExpressionH[FloatH] {
   override def resultType: ReferenceH[FloatH] = ReferenceH(ShareH, InlineH, ReadonlyH, FloatH())
 }
